+++ tests/input/mock_ncbi_objects.py
